Added methods to change profile picture & name in UserRepository.kt & created UserViewModel.kt to use those methods. These are currently only being used in profile settings.
Added methods in Extensions.kt for loading image from URI, and from a path of a file in Firebase Storage. Also added a method to convert a file path of firebase storage into storage reference.

Added a static method INSTANCE(..) in ProgressDialog.kt to get the instance. In future the constructor for ProgressDialog.kt will be made private and the INSTANCE(..) method will be the only way to get an instance. This is done to have only a single instance of ProgressDialog at any given time.

Functionality to reuse the current fragment on Screen rotation added to ActivityMain.kt.

Added functionality to preserve modifications in FragmentSetting.kt on screen rotation. Also, the methods to update profile data are removed from this fragment and the methods from UserViewModel.kt are being used.